[cronet_http] Upgrade jnigen to 0.9.2 to close #1213

+++ pkgs/cronet_http/lib/src/jni/jni_bindings.dart
@@ -749,7 +749,7 @@ class URL extends jni.JObject {
                       (
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>,
-                        ffi.Int64,
+                        ffi.Int32,
                         ffi.Pointer<ffi.Void>
                       )>)>>("globalEnv_NewObject")
       .asFunction<
@@ -831,7 +831,7 @@ class URL extends jni.JObject {
                       (
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>,
-                        ffi.Int64,
+                        ffi.Int32,
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>
                       )>)>>("globalEnv_NewObject")
@@ -1544,7 +1544,7 @@ class Executors extends jni.JObject {
   static final _newFixedThreadPool = ProtectedJniExtensions.lookup<
               ffi.NativeFunction<
                   jni.JniResult Function(ffi.Pointer<ffi.Void>,
-                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int64,)>)>>(
+                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int32,)>)>>(
           "globalEnv_CallStaticObjectMethod")
       .asFunction<
           jni.JniResult Function(
@@ -1568,7 +1568,7 @@ class Executors extends jni.JObject {
   static final _newWorkStealingPool = ProtectedJniExtensions.lookup<
               ffi.NativeFunction<
                   jni.JniResult Function(ffi.Pointer<ffi.Void>,
-                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int64,)>)>>(
+                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int32,)>)>>(
           "globalEnv_CallStaticObjectMethod")
       .asFunction<
           jni.JniResult Function(
@@ -1619,7 +1619,7 @@ class Executors extends jni.JObject {
                   jni.JniResult Function(
                       ffi.Pointer<ffi.Void>,
                       jni.JMethodIDPtr,
-                      ffi.VarArgs<(ffi.Int64, ffi.Pointer<ffi.Void>)>)>>(
+                      ffi.VarArgs<(ffi.Int32, ffi.Pointer<ffi.Void>)>)>>(
           "globalEnv_CallStaticObjectMethod")
       .asFunction<
           jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int,
@@ -1808,7 +1808,7 @@ class Executors extends jni.JObject {
   static final _newScheduledThreadPool = ProtectedJniExtensions.lookup<
               ffi.NativeFunction<
                   jni.JniResult Function(ffi.Pointer<ffi.Void>,
-                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int64,)>)>>(
+                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int32,)>)>>(
           "globalEnv_CallStaticObjectMethod")
       .asFunction<
           jni.JniResult Function(
@@ -1834,7 +1834,7 @@ class Executors extends jni.JObject {
                   jni.JniResult Function(
                       ffi.Pointer<ffi.Void>,
                       jni.JMethodIDPtr,
-                      ffi.VarArgs<(ffi.Int64, ffi.Pointer<ffi.Void>)>)>>(
+                      ffi.VarArgs<(ffi.Int32, ffi.Pointer<ffi.Void>)>)>>(
           "globalEnv_CallStaticObjectMethod")
       .asFunction<
           jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int,
@@ -2457,7 +2457,7 @@ class CronetEngine_Builder extends jni.JObject {
   static final _enableQuic = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallObjectMethod")
+                  ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_CallObjectMethod")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -2480,7 +2480,7 @@ class CronetEngine_Builder extends jni.JObject {
   static final _enableHttp2 = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallObjectMethod")
+                  ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_CallObjectMethod")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -2503,7 +2503,7 @@ class CronetEngine_Builder extends jni.JObject {
   static final _enableSdch = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallObjectMethod")
+                  ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_CallObjectMethod")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -2526,7 +2526,7 @@ class CronetEngine_Builder extends jni.JObject {
   static final _enableBrotli = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallObjectMethod")
+                  ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_CallObjectMethod")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -2549,7 +2549,7 @@ class CronetEngine_Builder extends jni.JObject {
   static final _enableHttpCache = ProtectedJniExtensions.lookup<
               ffi.NativeFunction<
                   jni.JniResult Function(ffi.Pointer<ffi.Void>,
-                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int64, ffi.Int64)>)>>(
+                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int32, ffi.Int64)>)>>(
           "globalEnv_CallObjectMethod")
       .asFunction<
           jni.JniResult Function(
@@ -2579,8 +2579,8 @@ class CronetEngine_Builder extends jni.JObject {
                   ffi.VarArgs<
                       (
                         ffi.Pointer<ffi.Void>,
-                        ffi.Int64,
-                        ffi.Int64
+                        ffi.Int32,
+                        ffi.Int32
                       )>)>>("globalEnv_CallObjectMethod")
       .asFunction<
           jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
@@ -2612,7 +2612,7 @@ class CronetEngine_Builder extends jni.JObject {
                       (
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>,
-                        ffi.Int64,
+                        ffi.Uint8,
                         ffi.Pointer<ffi.Void>
                       )>)>>("globalEnv_CallObjectMethod")
       .asFunction<
@@ -2654,7 +2654,7 @@ class CronetEngine_Builder extends jni.JObject {
                   jni.JniResult Function(
                       ffi.Pointer<ffi.Void>,
                       jni.JMethodIDPtr,
-                      ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallObjectMethod")
+                      ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_CallObjectMethod")
           .asFunction<
               jni.JniResult Function(
                   ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -2818,7 +2818,7 @@ class CronetEngine extends jni.JObject {
                   jni.JThrowablePtr Function(
                       ffi.Pointer<ffi.Void>,
                       jni.JMethodIDPtr,
-                      ffi.VarArgs<(ffi.Pointer<ffi.Void>, ffi.Int64)>)>>(
+                      ffi.VarArgs<(ffi.Pointer<ffi.Void>, ffi.Uint8)>)>>(
           "globalEnv_CallVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
@@ -3179,8 +3179,8 @@ class UploadDataProviders extends jni.JObject {
                   ffi.VarArgs<
                       (
                         ffi.Pointer<ffi.Void>,
-                        ffi.Int64,
-                        ffi.Int64
+                        ffi.Int32,
+                        ffi.Int32
                       )>)>>("globalEnv_CallStaticObjectMethod")
       .asFunction<
           jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
@@ -3393,7 +3393,7 @@ class UrlRequest_Builder extends jni.JObject {
   static final _setPriority = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallObjectMethod")
+                  ffi.VarArgs<(ffi.Int32,)>)>>("globalEnv_CallObjectMethod")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -3927,7 +3927,7 @@ class UrlRequest_StatusListener extends jni.JObject {
               jni.JThrowablePtr Function(
                   ffi.Pointer<ffi.Void>,
                   jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallVoidMethod")
+                  ffi.VarArgs<(ffi.Int32,)>)>>("globalEnv_CallVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
